requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
httpx==0.28.1